post rfc100 cleanup
