--- import-scripts/tempo-environment.sh
@@ -1,7 +1,0 @@
-DATABRICKS_PORT=443
-DATABRICKS_PATH=
-DATABRICKS_HOST=
-DATABRICKS_CATALOG=cdsi_public
-DATABRICKS_SCHEMA=tempo
-DATABRICKS_TOKEN=
-STUDY_DIRECTORY=
